requirements.txt
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyGithub==1.54.1
PyGObject==3.36.0
python-dotenv==0.15.0
python-gettext==4.0
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
Werkzeug==1.0.1
